chore: renamed get open issues using kebob case, added build utils to copy content images to public folder, fixed linting errors iwth single quotes within text in content pages
