+++ app/about/us/page.tsx
@@ -21,7 +21,7 @@ interface peopleTypes{
 
 const folderContent = await getAllContent('/about')
 const pageContent = folderContent.find(item => item.slug === 'us')
-function imagePath(imageName: string) {return path.join('/content/images/people/', imageName)}
+function imagePath(imageName: string) {return path.join('/images/people', imageName)}
 
 export default async function AboutUs() {
     return (
@@ -49,7 +49,7 @@ export default async function AboutUs() {
           <Card className="w-full border-none shadow-none rounded-none">
             <CardContent className="max-w-[1440px] mx-auto max-h-[600px] flex items-center px-6 py-10">
               <p className="text-black text-xl">
-                The Center for Computation and Visualization (CCV) is a center within the University's central IT organization, which is the <a href="https://it.brown.edu" target="_blank" rel="noopener noreferrer">Office of Information Technology (OIT)</a>. In addition to building and maintaining the University's <a href="https://it.brown.edu/tools-services" target="_blank" rel="noopener noreferrer">hundreds of enterprise software, systems, and hardware</a>, OIT is also responsible for driving the technological progress that enables scientific research. Executing on the University's research mission is the key role that CCV plays in OIT.
+                The Center for Computation and Visualization (CCV) is a center within the University&apos;s central IT organization, which is the <a href="https://it.brown.edu" target="_blank" rel="noopener noreferrer">Office of Information Technology (OIT)</a>. In addition to building and maintaining the University&apos;s <a href="https://it.brown.edu/tools-services" target="_blank" rel="noopener noreferrer">hundreds of enterprise software, systems, and hardware</a>, OIT is also responsible for driving the technological progress that enables scientific research. Executing on the University&apos;s research mission is the key role that CCV plays in OIT.
               </p>
             </CardContent>
           </Card>
